refactor: Made the generation of the feed more modular.
The feed will now be generated from a twig template. Also it will now be
a valid RSS feed.

The RSS feed will now be accessible at ?feed.rss so it is clear a RSS
feed is requested.

In order to produce a valid RSS feed a description has to be provided.
So this has to be set in the config.yaml.